+++ video-client/src/main/java/com/hudren/homevideo/TitleActivity.java
@@ -18,8 +18,6 @@ public class TitleActivity
         extends VideoActivity
         implements IVideoActivity
 {
-    private Title title;
-    private Video video;
 
     @Override
     public void onCreate( Bundle savedInstanceState )
@@ -51,28 +49,6 @@ public class TitleActivity
         getMenuInflater().inflate( R.menu.title, menu );
 
         return super.onCreateOptionsMenu( menu );
-    }
-
-    @Override
-    public boolean onOptionsItemSelected( MenuItem item )
-    {
-        int id = item.getItemId();
-
-        switch ( id )
-        {
-        case R.id.action_play:
-            play( title, video );
-            return true;
-
-        case R.id.action_download:
-            List<Video> videos = new ArrayList<>();
-            videos.add( video );
-
-            startDownloading( videos );
-            return true;
-        }
-
-        return super.onOptionsItemSelected( item );
     }
 
     @Override