+++ CHANGES.md
@@ -1,6 +1,14 @@
 # Videos@Home Changelog
 
-## Version 0.1
+## Version 1.1.1
+
+* Added links for IMDb and Netflix
+
+## Version 1.1
+
+* Added support for titles w/multiple videos
+
+## Version 1.0
 
 * Initial working release